requirements.txt
ollama
langchain
langchain-text-splitters
langchain-chroma
langchain-ollama
speechrecognition
pyttsx3
lxml
pyaudio
python-dotenv
pillow
opencv-python
tensorflow
torch
torchvision
numpy
matplotlib
scikit-learn
scipy
python-dateutil
tqdm